Retain old versions of lambda layer
Without this, if a deployment which includes a new version of the layer
fails partway through, deployment of the dependent stacks cannot be
rolled-back.

Old versions of the lambda layer should be manually deleted once the
dependent stacks have all been upgraded to use the new layer version,
which happens automatically when they are re-deployed.