Pinpoint app, frontend user with acces, examples
The frontend should log all events we wish to track directly to the
pinpoint api. This is what Amplify does.

The direct cognito user pool integration is not useful to us, as all it
does is update pinpoint endpoints with a unique user id if the login
goes through the user pool. Hence it's not implemented here.

We don't send all our logins throught the user pool (we use the identity
pool for federated authentication) and it is trivial to update the
endpoint in pinpoint with the userId ourselves.

Part of #25