Improve google id token parsing
Decode the JWT locally instead of using network api. Allows for more
descriptive error messages and logging of why the token was invalid, as
well as being more performant.

Also, don't raise a server error if the token was invalid. Log a warning
and return a client error.

Includes a new entry in the AWS secrets manager to store our google
client app ids.