Per-user chat message notifications
Requires us to to the fan-out of message notificaitons in code, rather
than relying on appsync's subscription mechanisms.

This implementation of that fan out won't scale to huge group chats
(should be fine up to 100 or so users), but when we get up to larger
chats we will need to be smarter about how we sent notifications to
users.