Bug fix chat memberships
When creating a new chat, the first message (added in the same api call)
was sometimes failing today update membership activity timestampes
because the newly-created chat item in dynamo didn't show up (because of
eventual consistency).

This change removes the need to look up the chat item before updating
the membership activity timestamps.