Handle apple appstore secrets better
Serverless's 'ssm' variable syntax is tempting because it's so simple,
but it leaves the secrets laying around in plain text both on the
machine the deployment was done from and in the lambda function's
environment variables. Neither is really acceptable.

With this commit the backend will retrieve the appstore secret (and
non-secret params, for simplicity) at runtime, thus avoiding the above
security issues.